refactor & lots of error handling
