--- server/api/controllers/examples/router.ts
@@ -1,7 +1,0 @@
-import express from 'express'
-import controller from './controller'
-export default express
-  .Router()
-  .post('/', controller.create)
-  .get('/', controller.all)
-  .get('/:id', controller.byId)
