+++ server/api/filecoin/index.ts
@@ -1,0 +1,26 @@
+import express from 'express'
+import * as wallet from './wallet'
+import * as message from './message'
+import * as misc from './misc'
+
+export default express
+  .Router()
+
+  // wallet
+  .post('/wallet', wallet.create)
+  .get('/wallet', wallet.get)
+  .get('/address', wallet.getAggAddress)
+  .get('/balance/:address', wallet.getBalance)
+
+  // messages
+  .get('/format', message.format)
+  .post('/message', message.cosign)
+  .get('/message/:cid', message.getStatus)
+  .get('/waitmsg/:cid', message.waitForReceipt)
+  .get('/receipts/:publicKey', message.getPastReceipts)
+
+  // misc
+  .get('/provider/address', misc.getProviderAddress)
+  .get('/provider/balance/:publicKey', misc.getProviderBalance)
+  .get('/blockheight', misc.getBlockHeight)
+  .get('/did', misc.getServerDid)
